+++ Halogen/src/TimeManage.h
@@ -29,7 +29,7 @@ public:
 	~SearchTimeManage();
 
 	bool ContinueSearch();	//Should I search to another depth, or stop with what ive got?
-	bool AbortSearch(int nodes);		//should I attempt to stop searching right now? Nodes is passed because we only want to check the exact time every 1000 nodes or so
+	bool AbortSearch(uint64_t nodes);		//should I attempt to stop searching right now? Nodes is passed because we only want to check the exact time every 1000 nodes or so
 
 	void StartSearch(int ms);	//pass the allowed search time maximum in milliseconds
